Tried Restructing the project.
I observed that the td elements were not getting styled. Upon invesetigation, I found that they were recieving styles from the materialize.css file from the materialize css framework I am using.
To override the styles, I tried the !important css value. But somehow it was not working. I figured out I will get successful but the time it will take to get to the root of it will be huge.
So in order to avoid the wastage of time, I decided to focus on learning other things.

The main purpose of starting this project was to create it using the template element tag. In this regard, I did have some significant learnings.

THe other purpose was to create the project as a tool for using it in my office. I decided that I will create it later from scratch if needed.

Bye Bye.