Worked on resolving problem
I am facing a logical see-saw like problem and I'am not able to solve it using plain JavaScript. I tried the functional approach as well as the Object oriented approach. I think it can be solved using the frameworks like React. It seems strange that when plain old JS cannot solve the problem, how is it that frameworks are able to solve it?? Perhaps I am missing something/some concept of javascript.